push token model fix
