+++ core/android/build.gradle
@@ -55,7 +55,7 @@ android {
 dependencies{
     compileOnly("com.moengage:moe-android-sdk:12.2.04")
     compileOnly("com.moengage:inapp:6.3.0")
-    api("com.moengage:plugin-base:3.1.4-SNAPSHOT")
+    api("com.moengage:plugin-base:3.1.5-SNAPSHOT")
     implementation "org.jetbrains.kotlin:kotlin-stdlib:$kotlin_version"
 }
 apply from: file("./user-agent.gradle")
