+++ core/lib/model/push/push_campaign_data.dart
@@ -8,4 +8,16 @@ class PushCampaignData {
   PushCampaign pushCampaign;
 
   PushCampaignData(this.platform, this.accountMeta, this.pushCampaign);
+
+  @override
+  String toString() {
+    return "{\n" +
+        "platform: ${platform.asString}" +
+        "\n" +
+        "accountMeta: ${accountMeta.toString()}" +
+        "\n" +
+        "pushCampaign: ${pushCampaign.toString()}" +
+        "\n" +
+        "}";
+  }
 }
