+++ core/lib/model/push/push_token_data.dart
@@ -1,9 +1,22 @@
 import 'package:moengage_flutter/model/platforms.dart';
-import 'package:moengage_flutter/model/push/push_token.dart';
+import 'moe_push_service.dart';
 
 class PushTokenData {
   Platforms platform;
-  PushToken token;
+  String token;
+  MoEPushService pushService;
 
-  PushTokenData(this.platform, this.token);
+  PushTokenData(this.platform, this.token, this.pushService);
+
+  @override
+  String toString() {
+    return "{\n" +
+        "platform: $platform" +
+        "\n" +
+        "token: $token" +
+        "\n" +
+        "pushService: $pushService" +
+        "\n" +
+        "}";
+  }
 }
