+++ core/lib/push_payload_mapper.dart
@@ -5,7 +5,6 @@ import 'package:moengage_flutter/model/platforms.dart';
 import 'package:moengage_flutter/model/push/moe_push_service.dart';
 import 'package:moengage_flutter/model/push/push_campaign.dart';
 import 'package:moengage_flutter/model/push/push_campaign_data.dart';
-import 'package:moengage_flutter/model/push/push_token.dart';
 import 'package:moengage_flutter/model/push/push_token_data.dart';
 import 'package:moengage_flutter/utils.dart';
 
@@ -17,8 +16,8 @@ class PushPayloadMapper {
       Map<String, dynamic> tokenData = json.decode(methodCallArgs);
       return PushTokenData(
           PlatformsExtension.fromString(tokenData[keyPlatform]),
-          PushToken(tokenData[keyPushToken],
-              MoEPushServiceExtention.fromString(tokenData[keyPushService])));
+          tokenData[keyPushToken],
+          MoEPushServiceExtention.fromString(tokenData[keyPushService]));
     } catch (exception) {
       print("$_tag Error: pushTokenFromJson() : ${exception.toString()}");
     }
@@ -27,6 +26,7 @@ class PushPayloadMapper {
 
   PushCampaignData? pushCampaignFromJson(dynamic methodCallArgs) {
     try {
+      print("$_tag Error: pushCampaignFromJson() : $methodCallArgs");
       Map<String, dynamic> pushCampaignPayload = json.decode(methodCallArgs);
       Map<String, dynamic> campaignData = pushCampaignPayload[keyData];
       return PushCampaignData(
